refactor: move deepHexifyBI to codec package
