+++ packages/molecule/src/index.ts
@@ -1,4 +1,4 @@
 export { createParser } from "./nearley";
 export { createCodecMap } from "./codec";
-export { toMolTypeMap, deepNumerifyBI, deepHexifyBI } from "./utils";
+export { toMolTypeMap } from "./utils";
 export type { Parser, CodecMap, MolTypeMap, MolType } from "./type";
